add link user name and images's post to their profile, fix 50% of like, comment post
