+++ app/src/main/java/com/hellodoc/healthcaresystem/user/personal/OtherUserProfile.kt
@@ -1,552 +1,371 @@
-
 package com.hellodoc.healthcaresystem.user.personal
-
 import android.content.Context
+import android.content.SharedPreferences
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
-import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.LayoutDirection
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import androidx.navigation.NavHostController
-import coil.compose.rememberAsyncImagePainter
-import com.hellodoc.healthcaresystem.viewmodel.DoctorViewModel
-import com.hellodoc.healthcaresystem.responsemodel.GetDoctorResponse
-import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.PostColumn
-import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.ViewIntroduce
-import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.ViewRating
-import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.WriteReviewScreen
-import com.hellodoc.healthcaresystem.R
-import com.hellodoc.healthcaresystem.ui.theme.HealthCareSystemTheme
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.rememberNavController
-import com.auth0.android.jwt.JWT
+import coil.compose.AsyncImage
+import com.hellodoc.healthcaresystem.R
+import com.hellodoc.healthcaresystem.responsemodel.User
 import com.hellodoc.healthcaresystem.user.home.ZoomableImageDialog
 import com.hellodoc.healthcaresystem.user.home.booking.doctorId
+import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.InteractPostManager
+import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.PostColumn
+import com.hellodoc.healthcaresystem.user.post.userId
 import com.hellodoc.healthcaresystem.viewmodel.PostViewModel
+import com.hellodoc.healthcaresystem.viewmodel.UserViewModel
 
+
+@RequiresApi(Build.VERSION_CODES.O)
+@Preview(showBackground = true)
 @Composable
-fun UserInfoSkeleton() {
-    ConstraintLayout(
-        modifier = Modifier
-            .background(Color.Cyan)
-            .height(330.dp)
-            .fillMaxWidth()
-    ) {
-        val (imgPlaceholder, line1, line2) = createRefs()
+fun PreviewProfileOtherUserPage() {
+    // Fake SharedPreferences
+    val context = LocalContext.current
+    val sharedPreferences = context.getSharedPreferences("preview_prefs", Context.MODE_PRIVATE)
 
-        Box(
-            modifier = Modifier
-                .size(110.dp)
-                .clip(CircleShape)
-                .background(Color.LightGray)
-                .constrainAs(imgPlaceholder) {
-                    top.linkTo(parent.top, margin = 45.dp)
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-        )
-
-        Box(
-            modifier = Modifier
-                .height(24.dp)
-                .width(120.dp)
-                .background(Color.LightGray)
-                .constrainAs(line1) {
-                    top.linkTo(imgPlaceholder.bottom, margin = 15.dp)
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-        )
-
-        Box(
-            modifier = Modifier
-                .height(20.dp)
-                .width(180.dp)
-                .background(Color.LightGray)
-                .constrainAs(line2) {
-                    top.linkTo(line1.bottom, margin = 10.dp)
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-        )
-    }
+    // Fake NavController
+    val navController = rememberNavController()
+    ProfileOtherUserPage(
+        sharedPreferences = sharedPreferences,
+        navHostController = navController
+    )
 }
-
-var doctorID = ""
-
-var doctorName = ""
-
-var doctorAddress = ""
-
-var specialtyName = ""
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun ProfileScreen(navHostController: NavHostController) {
-    val sharedPreferences = navHostController.context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
-    val viewModel: DoctorViewModel = viewModel(factory = viewModelFactory {
-        initializer { DoctorViewModel(sharedPreferences) }
+fun ProfileOtherUserPage(
+    sharedPreferences: SharedPreferences,
+    navHostController: NavHostController
+) {
+
+    // Khởi tạo ViewModel bằng custom factory để truyền SharedPreferences
+    val userViewModel: UserViewModel = viewModel(factory = viewModelFactory {
+        initializer { UserViewModel(sharedPreferences) }
     })
 
-    var selectedTab by remember { mutableIntStateOf(0) }
-    val showWriteReviewScreen = remember { mutableStateOf(false) }
-
+    val postViewModel: PostViewModel = viewModel(factory = viewModelFactory {
+        initializer { PostViewModel(sharedPreferences) }
+    })
+    val posts by postViewModel.posts.collectAsState()
+//    val comments by postViewModel.comments.collectAsState()
+    var shouldReloadPosts by remember { mutableStateOf(false) }
+    val navEntry = navHostController.currentBackStackEntry
+    val reloadTrigger = navEntry?.savedStateHandle?.getLiveData<Boolean>("shouldReload")?.observeAsState()
     val savedStateHandle = navHostController.previousBackStackEntry?.savedStateHandle
+    var thisUserId =""
 
     LaunchedEffect(Unit) {
-        savedStateHandle?.get<String>("doctorId")?.let {
-            doctorId = it
+        savedStateHandle?.get<String>("UserId")?.let {
+            thisUserId = it
         }
-        savedStateHandle?.remove<String>("doctorId")
-
-        selectedTab = savedStateHandle?.get<Int>("selectedTab") ?: 0
-        savedStateHandle?.remove<Int>("selectedTab")
-        //viewModel.fetchDoctorById(doctorId)
+        userModel = if (userViewModel.getUserAttributeString("role") == "user") "User" else "Doctor"
+    }
+    LaunchedEffect(reloadTrigger?.value) {
+        if (reloadTrigger?.value == true) {
+            postViewModel.getAllPosts() // gọi lại danh sách mới
+            navHostController.currentBackStackEntry
+                ?.savedStateHandle?.set("shouldReload", false)
+        }
+    }
+    // Gọi API để fetch user từ server
+    LaunchedEffect(thisUserId, shouldReloadPosts) {
+        if (thisUserId.isNotEmpty()) {
+            userViewModel.getUser(thisUserId)
+            postViewModel.getPostByUserId(thisUserId)
+        }
     }
 
+    // Lấy dữ liệu user từ StateFlow
+    val user by userViewModel.user.collectAsState()
+    // Nếu chưa có user (null) thì không hiển thị giao diện
 
-//    LaunchedEffect(doctorId) {
-//        doctorId?.let { viewModel.fetchDoctorById(it) }
-//    }
+    val context = LocalContext.current
+    var reportedPostId by remember { mutableStateOf<String?>(null) }
+    var showReportDialog by remember { mutableStateOf(false) }
 
-    LaunchedEffect(doctorId) {
-        doctorId.let { viewModel.fetchDoctorWithStats(it) }
-    }
+    println("user lấy ra đc: "+ user)
+    if (user==null) return
 
-    val doctor by viewModel.doctor.collectAsState()
-    val isLoading by viewModel.isLoading.collectAsState()
     var selectedImageUrl by remember { mutableStateOf<String?>(null) }
-
-     doctorID = doctor?.id ?: ""
-
-     doctorName = doctor?.name ?: ""
-
-     doctorAddress = doctor?.address ?: ""
-
-     specialtyName = doctor?.specialty?.name ?: ""
-
-//    if (doctor != null) {
-//        println("doctorId" + doctor!!.id)
-//    };
-//    if (doctor != null) {
-//        println("doctorName" + doctor!!.name)
-//    };
-//    if (doctor != null) {
-//        println("doctorAddress" + doctor!!.address)
-//    };
-//    if (doctor != null) {
-//        println("specialtyName" + doctor!!.specialty.name)
-//    };
-
     if (selectedImageUrl != null) {
         ZoomableImageDialog(selectedImageUrl = selectedImageUrl, onDismiss = { selectedImageUrl = null })
     }
-    Scaffold(
-        bottomBar = {
-            if (!showWriteReviewScreen.value) {
-                when (selectedTab) {
-                    0 -> BookingButton(navHostController)
-                    1 -> WriteReviewButton { showWriteReviewScreen.value = true }
+
+    var showFullScreenComment by remember { mutableStateOf(false) }
+    var selectedPostIdForComment by remember { mutableStateOf<String?>(null) }
+    var showReportBox by remember { mutableStateOf(false) }
+
+    // Nếu có user rồi thì hiển thị UI
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .pointerInput(Unit) {
+                detectTapGestures {
+                    postViewModel.closeAllPostMenus()  //tắt menu post
+                    showReportBox = false
                 }
             }
-        }
-    ) { innerPadding ->
-        LazyColumn(
-            modifier = Modifier.padding(
-                start = innerPadding.calculateStartPadding(LayoutDirection.Ltr),
-                end = innerPadding.calculateEndPadding(LayoutDirection.Ltr),
-                bottom = innerPadding.calculateBottomPadding()
-            )
-        ) {
+    ) {
+        LazyColumn(modifier = Modifier.fillMaxSize()) {
             item {
-                if (isLoading) {
-                    UserInfoSkeleton()
-                } else {
-                    UserInfo(
-                        doctor = doctor,
-                        navHostController = navHostController,
-                        onImageClick = { selectedImageUrl = it}
-                    )
-                }
-            }
-            item {
-                OtherUserListScreen(
+                OtherProfileSection(
                     navHostController = navHostController,
-                    doctor = doctor,
-                    selectedTab = selectedTab,
-                    onTabSelected = { selectedTab = it },
-                    showWriteReviewScreen = showWriteReviewScreen,
-                    onImageClick = { selectedImageUrl = it}
+                    user = user!!,
+                    onClickShowReport = { showReportDialog = true },
+                    onImageClick = { selectedImageUrl = it },
+                    showReportBox = showReportBox,
+                    onToggleReportBox = { showReportBox = !showReportBox }
                 )
             }
+            item {
+                PostColumn(
+                    posts = posts,
+                    postViewModel = postViewModel,
+                    userId = thisUserId ?: "",
+                    navController = navHostController,
+                    onClickReport = { postId ->
+                        reportedPostId = postId
+                        showReportDialog = true
+                    },
+                    onShowComment = { postId ->
+                        selectedPostIdForComment = postId
+                        showFullScreenComment = true
+                    }
+                )
+            }
+        }
+        if ((showFullScreenComment && selectedPostIdForComment != null) ||
+            (showReportDialog && user != null)
+        ) {
+            print("vo duoc interact post manager")
+            InteractPostManager(
+                navHostController,
+                user,
+                postViewModel,
+                reportedPostId,
+                context,
+                showFullScreenComment,
+                selectedPostIdForComment,
+                showReportDialog,
+                onCloseComment = {
+                    showFullScreenComment = false
+                },
+                onHideReportDialog = {
+                    showReportDialog = false
+                }
+            )
+        }
+    }
+
+}
+
+@Composable
+fun OtherProfileSection(
+    navHostController: NavHostController,
+    user: User, onClickShowReport: () -> Unit,
+    onImageClick: (String) -> Unit,
+    showReportBox: Boolean,
+    onToggleReportBox: () -> Unit
+)
+{
+    Column(
+        modifier = Modifier.background(Color.Cyan)
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(10.dp)
+        ) {
+            OtherUserIntroSection(
+                user = user,
+                onClickShowReport = onClickShowReport,
+                navController = navHostController,
+                onImageClick = onImageClick,
+                showReportBox = showReportBox,
+                onToggleReportBox = onToggleReportBox
+            )
+            Spacer(modifier = Modifier.height(26.dp))
+            Spacer(modifier = Modifier.height(10.dp))
         }
     }
 }
 
 @Composable
-fun UserInfo(
-    doctor: GetDoctorResponse?,
-    navHostController: NavHostController,
+fun OtherUserIntroSection(
+    user: User,
+    onClickShowReport: () -> Unit,
+    navController: NavHostController,
     onImageClick: (String) -> Unit,
-    modifier: Modifier = Modifier
+    showReportBox: Boolean,
+    onToggleReportBox: () -> Unit
 ) {
-    ConstraintLayout(
-        modifier = modifier
-            .background(Color.Cyan)
-            .height(330.dp)
+    Box(
+        modifier = Modifier
             .fillMaxWidth()
+            .height(250.dp)
+            .padding(16.dp)
     ) {
-        val (imgIcon, backIcon, tvTitle, tvName, tvNFollower, tvFollowers, tvNFollowing, tvFollowing, tvNLike, tvLikes) = createRefs()
-
-        val imageUrl = doctor?.avatarURL ?: ""
-        val name = doctor?.name ?: "Tên bác sĩ"
-        val experience = doctor?.experience?.toString() ?: "69"
-        val patientsCount = doctor?.patientsCount?.toString() ?: "0"
-        val ratingsCount = doctor?.ratingsCount?.toString() ?: "0"
-
-        // Ảnh bác sĩ
-        Image(
-            painter = rememberAsyncImagePainter(model = imageUrl),
-            contentDescription = "Doctor Avatar",
-            modifier = Modifier
-                .clip(CircleShape)
-                .size(110.dp)
-                .constrainAs(imgIcon) {
-                    top.linkTo(parent.top, margin = 45.dp)
-                    start.linkTo(parent.start)
-                    end.linkTo(parent.end)
-                }
-                .clickable {
-                    onImageClick(imageUrl)
-                },
-            contentScale = ContentScale.Crop
-        )
-
-        Text(
-            text = "Bác sĩ",
-            fontSize = 24.sp,
-            fontWeight = FontWeight.Bold,
-            color = Color.Black,
-            modifier = Modifier.constrainAs(tvTitle) {
-                top.linkTo(imgIcon.bottom, margin = 15.dp)
-                start.linkTo(parent.start)
-                end.linkTo(parent.end)
-            }
-        )
-
-        Text(
-            text = name,
-            fontSize = 22.sp,
-            fontWeight = FontWeight.SemiBold,
-            color = Color.Black,
-            modifier = Modifier.constrainAs(tvName) {
-                top.linkTo(tvTitle.bottom, margin = 10.dp)
-                start.linkTo(parent.start)
-                end.linkTo(parent.end)
-            }
-        )
-
-        val verticalGuideLine30Start = createGuidelineFromStart(0.3f)
-        val verticalGuideLine30End = createGuidelineFromEnd(0.3f)
-        val horizontalGuideLine20Bot = createGuidelineFromBottom(0.2f)
-
-        Text(
-            text = "$experience năm",
-            style = TextStyle(fontWeight = FontWeight.SemiBold, fontSize = 26.sp, color = Color.Blue),
-            modifier = Modifier.constrainAs(tvNFollower) {
-                top.linkTo(horizontalGuideLine20Bot)
-                end.linkTo(verticalGuideLine30Start)
-            }
-        )
-
-        Text(
-            text = "Kinh nghiệm",
-            style = TextStyle(fontWeight = FontWeight.Medium, fontSize = 15.sp, color = Color.Black),
-            modifier = Modifier.constrainAs(tvFollowers) {
-                top.linkTo(tvNFollower.bottom, margin = 5.dp)
-                end.linkTo(verticalGuideLine30Start)
-            }
-        )
-
-        Text(
-            text = patientsCount,
-            style = TextStyle(fontWeight = FontWeight.SemiBold, fontSize = 26.sp, color = Color.Blue),
-            modifier = Modifier.constrainAs(tvNFollowing) {
-                top.linkTo(horizontalGuideLine20Bot)
-                start.linkTo(parent.start)
-                end.linkTo(parent.end)
-            }
-        )
-
-        Text(
-            text = "Bệnh nhân",
-            style = TextStyle(fontWeight = FontWeight.Medium, fontSize = 15.sp, color = Color.Black),
-            modifier = Modifier.constrainAs(tvFollowing) {
-                top.linkTo(tvNFollowing.bottom, margin = 5.dp)
-                start.linkTo(parent.start)
-                end.linkTo(parent.end)
-            }
-        )
-
-        Text(
-            text = ratingsCount,
-            style = TextStyle(fontWeight = FontWeight.SemiBold, fontSize = 26.sp, color = Color.Blue),
-            modifier = Modifier.constrainAs(tvNLike) {
-                top.linkTo(horizontalGuideLine20Bot)
-                start.linkTo(verticalGuideLine30End, margin = 20.dp)
-            }
-        )
-
-        Text(
-            text = "Đánh giá",
-            style = TextStyle(fontWeight = FontWeight.Medium, fontSize = 15.sp, color = Color.Black),
-            modifier = Modifier.constrainAs(tvLikes) {
-                top.linkTo(tvNLike.bottom, margin = 5.dp)
-                start.linkTo(verticalGuideLine30End, margin = 10.dp)
-            }
-        )
-
         Image(
             painter = painterResource(id = R.drawable.arrow_back),
             contentDescription = "Back Icon",
             modifier = Modifier
                 .clip(CircleShape)
                 .size(40.dp)
-                .clickable { navHostController.popBackStack() }
-                .constrainAs(backIcon) {
-                    top.linkTo(parent.top, margin = 16.dp)
-                    start.linkTo(parent.start, margin = 16.dp)
-                },
+                .clickable { navController.popBackStack() },
             contentScale = ContentScale.Crop
         )
-    }
-}
-
-
-@RequiresApi(Build.VERSION_CODES.O)
-@Composable
-fun OtherUserListScreen(
-    navHostController: NavHostController,
-    doctor: GetDoctorResponse?,
-    selectedTab: Int,
-    onTabSelected: (Int) -> Unit,
-    showWriteReviewScreen: MutableState<Boolean>,
-    onImageClick: (String) -> Unit
-) {
-    println("Doctor lay duoc: "+doctor)
-
-    val tabs = listOf("Thông tin", "Đánh giá", "Bài viết")
-    val context = LocalContext.current
-    val sharedPreferences = remember {
-        context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
-    }
-    val token = remember { sharedPreferences.getString("access_token", null) }
-
-    val postViewModel: PostViewModel = viewModel(factory = viewModelFactory {
-        initializer { PostViewModel(sharedPreferences) }
-    })
-    val posts by postViewModel.posts.collectAsState()
-
-    val jwt = remember(token) {
-        runCatching { JWT(token ?: throw IllegalArgumentException("Token is null")) }
-            .onFailure { it.printStackTrace() }
-            .getOrNull()
-    }
-
-    val currentUserId = remember(jwt) {
-        jwt?.getClaim("userId")?.asString() ?: ""
-    }
-
-    var refreshReviewsTrigger by rememberSaveable { mutableStateOf(false) }
-    var editingReviewId by remember { mutableStateOf<String?>(null) }
-    var editingRating by remember { mutableStateOf<Int?>(null) }
-    var editingComment by remember { mutableStateOf<String?>(null) }
-
-    var reportedPostId by remember { mutableStateOf<String?>(null) }
-    var showReportDialog by remember { mutableStateOf(false) }
-    var showFullScreenComment by remember { mutableStateOf(false) }
-    var selectedPostIdForComment by remember { mutableStateOf<String?>(null) }
-    LaunchedEffect(doctor?.id) {
-        doctor?.id?.let {
-            postViewModel.getPostByUserId(it)
-        }
-    }
-
-    var selectedImageUrl by remember { mutableStateOf<String?>(null) }
-    if (selectedImageUrl != null) {
-        ZoomableImageDialog(selectedImageUrl = selectedImageUrl, onDismiss = { selectedImageUrl = null })
-    }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White)
-    ) {
-        TabRow(
-            selectedTabIndex = selectedTab,
-            containerColor = Color.Cyan,
-            contentColor = Color.Black
+        Column(
+            modifier = Modifier.align(Alignment.Center),
+            horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            tabs.forEachIndexed { index, title ->
-                Tab(
-                    selected = selectedTab == index,
-                    onClick = { onTabSelected(index) },
-                    text = {
-                        Text(
-                            text = title,
-                            fontWeight = if (selectedTab == index) FontWeight.Bold else FontWeight.Normal
-                        )
-                    }
+            Spacer(modifier = Modifier.height(20.dp))
+            // Avatar
+            Box(
+                modifier = Modifier
+                    .size(120.dp)
+                    .clip(CircleShape)
+                    .background(Color.Gray.copy(alpha = 0.2f))
+            ) {
+                AsyncImage(
+                    model = user.avatarURL,
+                    contentDescription = "Avatar",
+                                                                                                                                                                                                                                                                                                                                                                                modifier = Modifier
+                        .fillMaxSize()
+                        .clip(CircleShape)
+                        .clickable {
+                            onImageClick(user.avatarURL)
+                        },
+                    contentScale = ContentScale.Crop,
                 )
             }
-        }
-        println("Vao dươc toi trang user khác")
-            when (selectedTab) {
-                0 -> ViewIntroduce(doctor = doctor,onImageClick)
-                1 -> {
-                    if (showWriteReviewScreen.value) {
-                        WriteReviewScreen(
-                            doctorId = doctor?.id ?: "",
-                            userId = currentUserId,
-                            initialRating = editingRating,
-                            initialComment = editingComment,
-                            reviewId = editingReviewId,
-                            onBackClick = {
-                                showWriteReviewScreen.value = false
-                                editingReviewId = null
-                                editingRating = null
-                                editingComment = null
-                            },
-                            onSubmitClick = { _, _ ->
-                                refreshReviewsTrigger = !refreshReviewsTrigger
-                                showWriteReviewScreen.value = false
-                                editingReviewId = null
-                                editingRating = null
-                                editingComment = null
-                            }
-                        )
-                    } else {
-                        ViewRating(
-                            doctorId = doctor?.id ?: "",
-                            refreshTrigger = refreshReviewsTrigger,
-                            onEditReview = { reviewId, rating, comment ->
-                                editingReviewId = reviewId
-                                editingRating = rating
-                                editingComment = comment
-                                showWriteReviewScreen.value = true
-                            },
-                            onDeleteReview = {
-                                refreshReviewsTrigger = !refreshReviewsTrigger
-                            }
-                        )
-                    }
-                }
 
-            2 -> PostColumn(
-                posts = posts,
-                postViewModel = postViewModel,
-                userId = com.hellodoc.healthcaresystem.user.post.userId ?: "",
-                navController =  navHostController,
-                onClickReport = { postId ->
-                    reportedPostId = postId
-                    showReportDialog = true
-                },
-                onShowComment = { postId ->
-                    selectedPostIdForComment = postId
-                    showFullScreenComment = true
-                }
-            )
-        }
-    }
-}
-
-@Composable
-fun BookingButton(navController: NavHostController) {
-
-    Box(
-        Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 10.dp)
-    ) {
-
-        Button(
-            onClick = {
-                navController.currentBackStackEntry?.savedStateHandle?.apply {
-                    set("doctorId", doctorID)
-                    set("doctorName", doctorName)
-                    set("doctorAddress", doctorAddress)
-                    set("specialtyName", specialtyName)
-                }
-                navController.navigate("booking")
-            },
-            colors = ButtonDefaults.buttonColors(containerColor = Color.Cyan),
-            shape = RoundedCornerShape(20.dp),
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(55.dp)
-                .align(Alignment.Center)
-        ) {
+            Spacer(modifier = Modifier.height(20.dp))
             Text(
-                text = "Đặt khám",
-                fontSize = 24.sp,
-                fontWeight = FontWeight.Medium,
-                modifier = Modifier.padding(vertical = 8.dp)
+                user.name,
+                fontWeight = FontWeight.Bold,
+                fontSize = 20.sp,
+                color = Color.Black
+            )
+            Spacer(modifier = Modifier.height(10.dp))
+            Text(
+                user.email,
+                fontSize = 14.sp,
+                color = Color.DarkGray
             )
         }
-    }
-}
 
-@Composable
-fun WriteReviewButton(onClick: () -> Unit) {
-    Box(
-        Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 10.dp)
-    ) {
-        Button(
-            onClick = onClick,
-            colors = ButtonDefaults.buttonColors(containerColor = Color.Black),
-            shape = RoundedCornerShape(12.dp),
+        // Icon 3 chấm
+        IconButton(
+            onClick = { onToggleReportBox() },
             modifier = Modifier
-                .fillMaxWidth()
-                .height(50.dp)
-                .align(Alignment.Center)
+                .align(Alignment.TopEnd)
+                .padding(8.dp)
         ) {
-            Text("Viết đánh giá", fontSize = 16.sp, color = Color.White)
+            Icon(
+                painter = painterResource(id = R.drawable.ic_more),
+                contentDescription = "Menu",
+                tint = Color.Black
+            )
         }
-    }
-}
 
-@RequiresApi(Build.VERSION_CODES.O)
-@Preview(showBackground = true, showSystemUi = true)
-@Composable
-fun OtherUserProfilePreview() {
-    val navController = rememberNavController()
-    HealthCareSystemTheme {
-        ProfileScreen(navController)
+        // Report Box
+        if (showReportBox) {
+            Column(
+                modifier = Modifier
+                    .align(Alignment.TopEnd)
+                    .padding(top = 48.dp, end = 8.dp)
+                    .width(250.dp)
+                    .background(Color.White, shape = RoundedCornerShape(8.dp))
+                    .border(1.dp, Color.LightGray, RoundedCornerShape(8.dp))
+                    .shadow(4.dp, RoundedCornerShape(8.dp))
+                    .padding(12.dp)
+            ) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .clickable {
+                            onToggleReportBox()
+                            onClickShowReport()
+                        }
+                ) {
+                    Text("Tố cáo & Báo lỗi", fontWeight = FontWeight.Bold)
+                    Spacer(modifier = Modifier.height(4.dp))
+                    Text("Phản ánh vi phạm hoặc lỗi hệ thống", fontSize = 13.sp)
+                }
+
+                Divider(
+                    color = Color.LightGray,
+                    thickness = 1.dp,
+                    modifier = Modifier.padding(vertical = 12.dp)
+                )
+
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .clickable {
+                            onToggleReportBox()
+                            navController.navigate("activity_manager")
+                        }
+                ) {
+                    Text("Quản lý hoạt động", fontWeight = FontWeight.Bold)
+                    Spacer(modifier = Modifier.height(4.dp))
+                    Text("Xem và kiểm soát các hoạt động", fontSize = 13.sp)
+                }
+            }
+        }
     }
 }
