+++ app/src/main/java/com/hellodoc/healthcaresystem/user/personal/UseOwnerProfile.kt
@@ -2,9 +2,7 @@ package com.hellodoc.healthcaresystem.user.personal
 import android.content.Context
 import android.content.SharedPreferences
 import android.os.Build
-import android.widget.Toast
 import androidx.annotation.RequiresApi
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -20,9 +18,7 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
@@ -30,9 +26,7 @@ import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Divider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -40,48 +34,35 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.window.Dialog
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import coil.compose.AsyncImage
-import coil.compose.rememberAsyncImagePainter
-import com.hellodoc.healthcaresystem.responsemodel.ContainerPost
-import com.hellodoc.healthcaresystem.responsemodel.ContentPost
-import com.hellodoc.healthcaresystem.responsemodel.FooterItem
 import com.hellodoc.healthcaresystem.R
-import com.hellodoc.healthcaresystem.requestmodel.ReportRequest
-import com.hellodoc.healthcaresystem.responsemodel.PostResponse
 import com.hellodoc.healthcaresystem.responsemodel.User
-import com.hellodoc.healthcaresystem.retrofit.RetrofitInstance
 import com.hellodoc.healthcaresystem.user.home.ZoomableImageDialog
-import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.FullScreenCommentUI
+import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.InteractPostManager
 import com.hellodoc.healthcaresystem.user.personal.otherusercolumn.PostColumn
 import com.hellodoc.healthcaresystem.user.post.userId
 import com.hellodoc.healthcaresystem.viewmodel.PostViewModel
 import com.hellodoc.healthcaresystem.viewmodel.UserViewModel
-import kotlinx.coroutines.launch
 
 var userId = ""
 var userModel = ""
@@ -102,6 +83,7 @@ fun PreviewProfileUserPage() {
     )
 }
 
+@RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun ProfileUserPage(
     sharedPreferences: SharedPreferences,
@@ -145,7 +127,6 @@ fun ProfileUserPage(
     val user by userViewModel.user.collectAsState()
     // Nếu chưa có user (null) thì không hiển thị giao diện
 
-    val coroutineScope = rememberCoroutineScope()
     val context = LocalContext.current
     var reportedPostId by remember { mutableStateOf<String?>(null) }
     var showReportDialog by remember { mutableStateOf(false) }
@@ -201,152 +182,25 @@ fun ProfileUserPage(
                 )
             }
         }
-        if (showFullScreenComment && selectedPostIdForComment != null) {
-            FullScreenCommentUI(
-                postId = selectedPostIdForComment!!,
-                onClose = { showFullScreenComment = false },
-                postViewModel = postViewModel,
-                currentUserId = userId
-            )
-        }
-
-        if (showReportDialog && user != null) {
-            var selectedType by remember { mutableStateOf("Ứng dụng") }
-            var reportContent by remember { mutableStateOf("") }
-
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color.Black.copy(alpha = 0.3f))
-                    .clickable(enabled = true, onClick = {}),
-                contentAlignment = Alignment.Center
-            ) {
-                Column(
-                    modifier = Modifier
-                        .width(320.dp)
-                        .background(Color.White, shape = RoundedCornerShape(12.dp))
-                        .border(1.dp, Color.Gray)
-                        .padding(16.dp)
-                ) {
-                    Text("Báo cáo người dùng", fontWeight = FontWeight.Bold, fontSize = 18.sp)
-                    Spacer(modifier = Modifier.height(8.dp))
-
-                    Text("Người báo cáo", fontWeight = FontWeight.Medium)
-                    Text(user!!.name, color = Color.DarkGray)
-
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text("Loại báo cáo", fontWeight = FontWeight.Medium)
-
-                    Row(verticalAlignment = Alignment.CenterVertically) {
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier
-                                .clickable { selectedType = "Bác sĩ" }
-                                .padding(end = 10.dp)
-                        ) {
-                            RadioButton(
-                                selected = selectedType == "Bác sĩ",
-                                onClick = null  // <- để dùng chung onClick bên ngoài
-                            )
-                            Text("Bác sĩ", modifier = Modifier.padding(start = 5.dp))
-                        }
-
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier
-                                .clickable { selectedType = "Ứng dụng" }
-                                .padding(end = 10.dp)
-                        ) {
-                            RadioButton(
-                                selected = selectedType == "Ứng dụng",
-                                onClick = null
-                            )
-                            Text("Ứng dụng", modifier = Modifier.padding(start = 5.dp))
-                        }
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier.clickable { selectedType = "Bài viết" }
-                        ) {
-                            RadioButton(
-                                selected = selectedType == "Bài viết",
-                                onClick = null
-                            )
-                            Text("Bài viết", modifier = Modifier.padding(start = 5.dp))
-                        }
-                    }
-
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text("Nội dung báo cáo", fontWeight = FontWeight.Medium)
-                    TextField(
-                        value = reportContent,
-                        onValueChange = { reportContent = it },
-                        placeholder = { Text("Nhập nội dung...") },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(120.dp)
-                    )
-
-                    Spacer(modifier = Modifier.height(12.dp))
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween
-                    ) {
-                        Text(
-                            "Huỷ",
-                            color = Color.Red,
-                            modifier = Modifier
-                                .clickable { showReportDialog = false }
-                                .padding(8.dp),
-                            fontWeight = FontWeight.Medium
-                        )
-
-                        Button(onClick = {
-                            coroutineScope.launch {
-                                try {
-                                    val model =
-                                        if (user!!.role.lowercase() == "doctor") "Doctor" else "User"
-                                    val response = RetrofitInstance.reportService.sendReport(
-                                        ReportRequest(
-                                            reporter = user!!.id,
-                                            reporterModel = model,
-                                            content = reportContent,
-                                            type = selectedType,
-                                            reportedId = user!!.id,
-                                            postId = if (selectedType == "Bài viết") reportedPostId else null
-                                        )
-                                    )
-
-                                    if (response.isSuccessful) {
-                                        Toast.makeText(
-                                            context,
-                                            "Đã gửi báo cáo thành công",
-                                            Toast.LENGTH_SHORT
-                                        ).show()
-                                        reportedPostId = null
-                                    } else {
-                                        println(response)
-                                        Toast.makeText(
-                                            context,
-                                            "Gửi báo cáo thất bại",
-                                            Toast.LENGTH_SHORT
-                                        ).show()
-                                    }
-                                } catch (e: Exception) {
-                                    Toast.makeText(
-                                        context,
-                                        "Lỗi kết nối đến server",
-                                        Toast.LENGTH_SHORT
-                                    ).show()
-                                    e.printStackTrace()
-                                }
-                            }
-                            showReportDialog = false
-                        }) {
-                            Text("Gửi báo cáo")
-                        }
-                    }
+        if ((showFullScreenComment && selectedPostIdForComment != null) ||
+            (showReportDialog && user != null)) {
+            print("vo duoc interact post manager")
+            InteractPostManager(
+                navHostController,
+                user,
+                postViewModel,
+                reportedPostId,
+                context,
+                showFullScreenComment,
+                selectedPostIdForComment,
+                showReportDialog,
+                onCloseComment = {
+                    showFullScreenComment = false
+                },
+                onHideReportDialog = {
+                    showReportDialog = false
                 }
-            }
+            )
         }
     }
 }
