+++ app/src/main/res/xml/network_security_config.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <network-security-config xmlns:android="http://schemas.android.com/apk/res/android">
     <domain-config cleartextTrafficPermitted="true">
-        <domain includeSubdomains="true">192.168.222.22</domain>
+        <domain includeSubdomains="true">192.168.0.102</domain>
         <domain includeSubdomains="true">192.168.100.5</domain>
         <domain includeSubdomains="true">192.168.1.5</domain>
         <domain includeSubdomains="true">192.168.1.9</domain>
